pyproject.toml
[project]
name = "inkygallery"
version = "0.1.0"
description = "Minimal backend-first image library and playback service for Inky displays."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1,<4",
    "waitress>=3.0,<4",
    "pillow>=11.0,<12",
    "requests>=2.32,<3",
    "psutil>=7.0,<8",
    "pytz>=2024.2,<2027",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
hardware = [
    "pi-heif>=0.22,<1",
    "inky>=2.1,<3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
